refactor(api-v3): move the Hub error mapper out of feedbackRecords
The taxonomy routes need the same Hub-error mapping, and a resource module must not import from a
sibling resource, so the generic half moves to app/api/v3/lib/hub-errors.ts.

A pure relocation: the caller-supplied 503 detail this depends on already landed on the base branch,
addressing the same review point for feedbackRecords. The two domain strings
(EMBEDDINGS_UNAVAILABLE_DETAIL, EMBEDDING_PENDING_DETAIL) stay behind with the surface that produces
them, and no call-site signature changes.

Its test moves with it, and stops borrowing a feedbackRecords constant as a fixture — any
caller-supplied string proves the same plumbing without the shared module's test reaching into a
feature module.
